Adds asyncMain option to init command
Addresses #65

I'm not proud of the approach I took to keep the JSON "pretty". I didn't want an empty defines object being output when asyncMain was not invoked, but if another defines option is added, this will become awkward.

Creating a JSON object and writing the stringified object to the filesystem might be more future proof than using strings and an ejs template.

Finally, the property name for XS' manifest.json is `async_main`, but ES Lint complains about that style so I've used `asyncMain` for the param. It bothers me that they don't match, but a contributor changing `.eslint.js` is kinda like having a houseguest rearrange furniture. I don't want to be that guy.